Setup Bugzilla during Stackato staging
A post-staging script (setup.pl) will download Bugzilla and configure it
with a minimal setup (for example.com domain).

CGI::Emulate::PSGI has a bug that prevents it to work with uWSGI
right now, so we are running Bugzilla under plackup for the time
being:  https://rt.cpan.org/Ticket/Display.html?id=72684

This setup includes the *minimal* set up of required modules and
does not install all optional ones.