List 'Plack' in requirements.txt
Previously only modules from the Plack package were listed
(redundantly), and only the first of them (Plack::Builder) was
actually known to the PPM index.  As long as Stackato installed all
dependencies in the order they are listed in requirements.txt this
works fine, because the PPM client then notices that
Plack::App::URLMap and Plack::App::WrapCGI were already installed by
the Plack::Builder dependency.

In Stackato 0.7 and later PPM will evaluate all requirements in
parallel, so the fact that P::A::URLMap and P::A::WrapCGI are not
included in the PPM database throws an error.

See also: http://bugs.activestate.com/show_bug.cgi?id=92110